A feature is changed about placing the wall
When we try to place a wall somewhere not free, we send the put to where it was before. But if that place is also occupied, we were removing from the board in the previous version. However, now it show the previous place for 0.5 seconds and then it wants the player to place it somewhere free.